Pass original mutation to subscribers on object-style commit
After object-style commit subscribers got back the original mutation
wrapped into a new mutation as payload. This was counter-intuitive
and prevented the reuse of mutation objects. For example after the call
commit({ type: 'TEST', amount: 2 }) one would expect to receive the
same object in the subscriber but actually
{ type: 'TEST', payload: { type: 'TEST', amount: 2 } })
was received with the original implementation.