Explained how to use mutations with Vue templates
When learning Vuex, I found this to be a little confusing that 'mutations' were not automatically available just like 'methods' were.  Given the way in which most things with Vue just work 'batteries included,' I think it's important to call this out that the developer must explicitly make mutations available through methods.