Fix false negative when checking `payload: false`
Here is my example code:
```javascript
    test('commits setContributionId', (done) => {
      const state = {
        comments: []
      }
      testAction(action, 42, state, [
        // ...
        { type: 'isLoading', payload: false}
      ], done)
    })
```
I'm passing `42` as an argument to `testAction` and I want to verify that `commit('isLoading', payload)` is executed. Here the payload is not the payload of the testAction helper (in my case 42) but the second argument of `commit`. That might be `false` and my case it is. If I change the payload of the last expected commit to be `true` no assertion error is raised. The fix is to make the if condition in the mocked commit method more specific.