Given a mutation payload, don't skip assertion
If a user specifies an expected payload, the assertion **must** be
executed.

Let's say your store action looks like this:

```
const somePayload = undefined
// this is the bug to be exposed, `somePayload` should not be undefined

commit('isLoading', somePayload)
```

And the developer defines an expected payload in the test:

```js
testAction(action, 42, state, [
  //...
  { type: 'isLoading', payload: false }
], done)
```

The test will unexpectedly pass as the assertion is skipped.
Instead we always want to check the assertion if any payload was
defined by the developer.