action subscribers would receive a call even when the action failed (promise.reject, js error)
This way the action subscribers of the 'after' would not miss the end of the action.
The true/false value might be useful for after action handlers, but the result might be too tempting to be used for modification for example, might be that too much info is handled out in this way. The same goes for the success-side of the after subscribers too.