Return rejected promise instead of throwing
Fixes #1969

This code was deliberately throwing an error for the sole purpose of re-rejecting a promise that had been caught in order to emit an event for the devtools hook. However, throwing an error will also trigger the native browser devtools' "pause on exception" behaviour because the error is being thrown without a `try..catch` block. In order to re-reject the promise safely we can use Promise.reject() - this gets handled by the promise in the same way as using `throw`, but the browser devtools will not pause.