Define state when preserved is true, but is nothing to preserve
Vuex handles modules defined in the constructor options different from those registered with store.registerModule. For registerModule calls, the modules pass for a variety of checks before being embedded in the global store.

When checking in "installModule" if the module should be at the root and if the "state" should not be preserved is not observed that if "hot" is set to true, but the state was not defined before throws an error inside the mutation called to change state like "Cannot read properties of undefined (reading 'foo')".

That occurs because it was preserved a no existing state. That's why we should check if the state already exists if so, it will be preserved otherwise we set the default from the module. 

The fast way to verify this bug is creating a project from scratch, adding vuex, creating a module (I did it in a separate file), setting the namespace to true, registering it like "store.registerModule('bar', module, {preserveState:true})". Now at a vue file call from a a method "this.$store.dispatch('bar/setValue', {name:'foo'})".

[#2083]